cmplete crud of note with http error

+++ src/controllers/notes.ts
@@ -1,14 +1,17 @@
 import e, {Request, RequestHandler, Response, NextFunction} from "express";
 import NoteModel from "../models/note";
 import app from "../app";
-import {text} from "node:stream/consumers";
 import createHttpError from "http-errors";
 import {throws} from "node:assert";
 import mongoose from "mongoose";
+import {PassThrough} from "node:stream";
 
 export const getNotes:RequestHandler = async (req: Request, res: Response,next:NextFunction) => {
     try {
         const notes = await NoteModel.find().exec();
+        if (!notes){
+            throw createHttpError(404,"No notes found")
+        }
         res.status(200).json(notes);
     } catch (error) {
         next(error)
@@ -17,8 +20,9 @@ export const getNotes:RequestHandler = async (req: Request, res: Response,next:N
 
 export const getNote: RequestHandler = async (req: Request, res: Response,next:NextFunction) => {
     // const  noteId  = req.params.noteId;
-    const {noteId} = req.params;
+    const {noteId} = req.params;       //this is very importanting name (noteId name)
     try {
+        //create a id is a mongoose id
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid note id")
         }
@@ -52,49 +56,45 @@ export const createNote: RequestHandler = async (req: Request, res: Response,nex
     }
 }
 
-// interface UpdateNoteParams {
-//     noteId: string;
-// }
-//
-// interface UpdateNoteBody {
-//     title?: string;
-//     text?: string;
-// }
-
-// export const updateNote: RequestHandler<UpdateNoteParams,unknown,UpdateNoteBody,unknown> = async (req, res) => {
-//     const {noteId} = req.params;
- //     const {newTitle, newText} = req.body;
-//     try {
-//         let note = await NoteModel.findById(noteId).exec();
-//         if (!note) {
-//             res.status(404).send("note not found");
-//         }
-//
-//         note.title = newText;
-//         note.text = newTitle;
-//         const updatedNote = await note?.save();
-//         res.status(200).send(updatedNote)
-//     } catch (e) {
-//         res.status(400).json({message: e.message});
-//     }
-// }
+export const updateNote: RequestHandler = async (req:Request, res:Response,next:NextFunction) => {
+    const {noteId} = req.params;
+     const {title, text} = req.body;
+    try {
+        if(!mongoose.isValidObjectId(noteId)){
+            throw createHttpError(400, "Invalid note id")
+        }
+        let note = await NoteModel.findById(noteId).exec();
+        if (!note) {
+            throw createHttpError(404,"note not found");
+        }
+        note.title = title;
+        note.text = text;
+        const updatedNote = await note?.save();
+        res.status(200).send(updatedNote)
+    } catch (error) {
+        next(error)
+    }
+}
 
 
 interface DeleteNoteParams {
     noteId: string;
 }
 
-export const deleteNote: (req: e.Request, res: e.Response) => Promise<void> = async (req: Request, res: Response) => {
+export const deleteNote=async (req:Request, res:Response,next:NextFunction)=> {
     const {noteId} = req.params;
     try {
+        if(!mongoose.isValidObjectId(noteId)){
+            throw createHttpError(400, "Invalid note id")
+        }
         const note = await NoteModel.findById(noteId).exec();
 
         if (!note) {
-            res.status(404).json({message: "Note not found"});
+            throw createHttpError(404,"note not found")
         }
         await note?.deleteOne();
         res.status(200).json({message: "Note deleted successfully"});
-    } catch (e) {
-        res.status(500).json({message: e.message});
+    } catch (error) {
+        next(error);
     }
 };
